feat: refactor smoke tests
