+++ examples/app.js
@@ -1,5 +1,8 @@
+import { fileURLToPath } from 'node:url';
 import { App, staticFiles } from '@limette/core';
+
+const root = fileURLToPath(new URL('.', import.meta.url));
 
 export const app = new App()
   .use(staticFiles)
-  .fsRoutes();
+  .fsRoutes({ vite: { root } });
