+++ package.json
@@ -14,6 +14,14 @@
   },
   "scripts": {
     "build": "tsdown",
+    "check": "deno check src/mod.ts src/vite/mod.ts src/deno.ts src/node.ts",
+    "check:examples": "deno check --config examples/deno.json examples/app.js examples/main.js examples/vite.config.ts",
+    "check:smoke": "deno check --config examples/deno.json tests/smoke/vite-build.ts tests/smoke/vite-server.ts tests/smoke/vite-assets.ts tests/smoke/vite-ssr.ts tests/smoke/vite-dev-routes.ts tests/smoke/vite-dev-server.ts tests/smoke/node-server.ts",
+    "test:examples": "npm --prefix examples run build",
+    "test:smoke": "npm run test:smoke:routes && deno run -A --config examples/deno.json tests/smoke/vite-build.ts && deno run -A --config examples/deno.json tests/smoke/vite-server.ts && deno run -A --config examples/deno.json tests/smoke/vite-assets.ts && deno run -A --config examples/deno.json tests/smoke/vite-ssr.ts",
+    "test:smoke:routes": "deno run -A npm:vite@^8.0.0 --config examples/vite.config.ts build --ssr ../tests/smoke/vite-routes.ts --outDir .vite-limette-smoke",
+    "test:smoke:server": "deno run -A --config examples/deno.json tests/smoke/node-server.ts && deno run -A --config examples/deno.json tests/smoke/vite-dev-routes.ts && deno run -A --config examples/deno.json tests/smoke/vite-dev-server.ts",
+    "pack:dry": "npm pack --dry-run",
     "prepack": "npm run build"
   },
   "exports": {
